requirements.txt
Flask
Flask-SQLAlchemy
pymysql
python-dotenv
flask-bcrypt